further built users router and consolidated CRUD functions into controllers
