+++ index.js
@@ -2,14 +2,16 @@ const express = require('express')
 const app = express()
 const dotenv = require('dotenv').config()
 const mongoose = require('mongoose')
-const User = require('./models/users')
 const bodyParser = require('body-parser')
+const usersRouter = require('./routers/users')
 
 //Connection to MongoDB
-mongoose.connect('mongodb://localhost:27017/blogapp', {
+
+mongoose.connect(process.env.URL, {
     useNewUrlParser: true,
     useUnifiedTopology: true
 })
+
 const db = mongoose.connection
 db.once('open', _ => {
     console.log('Database connected')
@@ -26,35 +28,7 @@ app.use(bodyParser.urlencoded({
 }))
 app.use(bodyParser.json())
 
-app.post('/users', (request, response, next) => {
-    const { firstName, lastName, email, password } = request.body
-    console.log(firstName, lastName, email, password)
-
-    if (!firstName || !lastName || !email || !password) {
-        return response.status(400).send('All fields required')
-    }
-
-    const userData = {
-        firstName: firstName,
-        lastName: lastName,
-        email: email,
-        password: password
-    }
-
-    User.create(userData, function (error, user) {
-        if (error) {
-            return response.send(error)
-        } else {
-            response.send(user)
-        }
-    })
-})
-
-app.get('/users', async (request, response) => {
-    const users = await User.find({})
-
-    return response.send(users)
-})
+app.use('/users', usersRouter)
 
 app.all('*', (request, response) => {
     response.sendStatus(404)
